moved var declarations out of ccr_test.h
