+++ include/ccr_test.h
@@ -13,10 +13,7 @@
 #include <assert.h>
 #include <stdlib.h>
 
-/* Err is used to keep track of errors within each set of tests,
- * total_err is the number of errors in the entire test program, which
- * generally cosists of several sets of tests. */
-static int total_err = 0, err = 0;
+#define MILLION 1000000
 
 /* This macro prints an error message with line number and name of
  * test program. */
