+++ test/tst_bzip2.c
@@ -31,6 +31,11 @@
 #define DEFLATE_LEVEL 3
 #define SIMPLE_VAR_NAME "data"
 
+/* Err is used to keep track of errors within each set of tests,
+ * total_err is the number of errors in the entire test program, which
+ * generally cosists of several sets of tests. */
+static int total_err = 0, err = 0;
+
 int
 main()
 {
