+++ test/tst_lz4.c
@@ -34,6 +34,11 @@
 #define NX_BIG 100
 #define NY_BIG 100
 
+/* Err is used to keep track of errors within each set of tests,
+ * total_err is the number of errors in the entire test program, which
+ * generally cosists of several sets of tests. */
+static int total_err = 0, err = 0;
+
 int
 main()
 {
