+++ test/tst_utils.c
@@ -1,0 +1,45 @@
+/*! \file Utility functions for tests.
+
+\internal
+Copyright 1993, 1994, 1995, 1996, 1997, 1998, 1999, 2000, 2001, 2002,
+2003, 2004, 2005, 2006, 2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014,
+2015, 2016, 2017, 2018
+University Corporation for Atmospheric Research/Unidata.
+
+See \ref copyright file for more info.
+
+
+*/
+
+#include "config.h"
+#include "ccr_test.h"
+#include <time.h>
+#include <sys/time.h>
+
+/** Subtract the `struct timeval' values X and Y, storing the result in
+   RESULT.  Return 1 if the difference is negative, otherwise 0.  This
+   function from the GNU documentation. */
+int
+nc4_timeval_subtract (result, x, y)
+   struct timeval *result, *x, *y;
+{
+   /* Perform the carry for the later subtraction by updating Y. */
+   if (x->tv_usec < y->tv_usec) {
+      int nsec = (y->tv_usec - x->tv_usec) / MILLION + 1;
+      y->tv_usec -= MILLION * nsec;
+      y->tv_sec += nsec;
+   }
+   if (x->tv_usec - y->tv_usec > MILLION) {
+      int nsec = (x->tv_usec - y->tv_usec) / MILLION;
+      y->tv_usec += MILLION * nsec;
+      y->tv_sec -= nsec;
+   }
+
+   /* Compute the time remaining to wait.
+      `tv_usec' is certainly positive. */
+   result->tv_sec = x->tv_sec - y->tv_sec;
+   result->tv_usec = x->tv_usec - y->tv_usec;
+
+   /* Return 1 if result is negative. */
+   return x->tv_sec < y->tv_sec;
+}
